Refactor sim_unicorn.cpp
- Moved structure and class definitions to a header file.
- Removed typedef for structs: In C++, we can declare structs without
  struct keyword.
- Converted macros to constants.
- Replaced C headers with C++ headers.
